Fixes style checks for main code
