+++ src/main/java/seedu/address/logic/parser/appointments/MissAppCommandParser.java
@@ -1,9 +1,10 @@
 //@@author woon17
-package seedu.address.logic.parser;
+package seedu.address.logic.parser.appointments;
 
 import static seedu.address.commons.core.Messages.MESSAGE_INVALID_COMMAND_FORMAT;
 
 import seedu.address.logic.commands.appointments.MissAppCommand;
+import seedu.address.logic.parser.Parser;
 import seedu.address.logic.parser.exceptions.ParseException;
