+++ src/main/java/seedu/address/logic/parser/queue/AddConsultationRoomCommandParser.java
@@ -1,4 +1,4 @@
-package seedu.address.logic.parser;
+package seedu.address.logic.parser.queue;
 
 import static seedu.address.commons.core.Messages.MESSAGE_INVALID_COMMAND_FORMAT;
 
@@ -6,11 +6,14 @@ import java.util.List;
 import java.util.stream.Stream;
 
 import seedu.address.commons.core.index.Index;
+import seedu.address.logic.commands.common.ReversibleActionPairCommand;
 import seedu.address.logic.commands.queue.AddConsultationRoomCommand;
 import seedu.address.logic.commands.queue.RemoveRoomCommand;
-import seedu.address.logic.commands.common.ReversibleActionPairCommand;
+import seedu.address.logic.parser.ArgumentMultimap;
+import seedu.address.logic.parser.Parser;
+import seedu.address.logic.parser.ParserUtil;
+import seedu.address.logic.parser.Prefix;
 import seedu.address.logic.parser.exceptions.ParseException;
-
 import seedu.address.model.Model;
 import seedu.address.model.ReferenceId;
 import seedu.address.model.person.Person;
