+++ src/main/java/seedu/address/logic/parser/queue/BreakCommandParser.java
@@ -1,36 +1,32 @@
-package seedu.address.logic.parser;
+package seedu.address.logic.parser.queue;
 
 import static seedu.address.commons.core.Messages.MESSAGE_INVALID_COMMAND_FORMAT;
 
-import java.util.Optional;
-
 import javafx.collections.ObservableList;
-
 import seedu.address.commons.core.index.Index;
 import seedu.address.logic.commands.NextCommand;
-import seedu.address.logic.commands.queue.UndoNextCommand;
 import seedu.address.logic.commands.common.ReversibleActionPairCommand;
+import seedu.address.logic.commands.queue.BreakCommand;
+import seedu.address.logic.commands.queue.ResumeCommand;
+import seedu.address.logic.parser.Parser;
+import seedu.address.logic.parser.ParserUtil;
 import seedu.address.logic.parser.exceptions.ParseException;
 import seedu.address.model.Model;
-import seedu.address.model.ReferenceId;
 import seedu.address.model.queue.Room;
 
 /**
  * Parses input arguments and creates a new DeleteCommand object
  */
-public class NextCommandParser implements Parser<ReversibleActionPairCommand> {
+public class BreakCommandParser implements Parser<ReversibleActionPairCommand> {
     public static final String MESSAGE_INVALID_INDEX = "Invalid index given";
-    public static final String MESSAGE_NO_PATIENT = "There are no patients in the queue";
 
     private Model model;
     private ObservableList<Room> filteredRoomList;
-    private ObservableList<ReferenceId> queueList;
     private Index index;
 
 
-    public NextCommandParser(Model model) {
+    public BreakCommandParser(Model model) {
         this.model = model;
-        this.queueList = model.getQueueList();
     }
 
     /**
@@ -49,14 +45,11 @@ public class NextCommandParser implements Parser<ReversibleActionPairCommand> {
         filteredRoomList = model.getConsultationRoomList();
         if (filteredRoomList.size() < index.getOneBased()) {
             throw new ParseException(MESSAGE_INVALID_INDEX);
-        } else if (queueList.size() == 0) {
-            throw new ParseException(MESSAGE_NO_PATIENT);
         }
-        ReferenceId patientBeingServed = queueList.get(0);
         Room roomToEdit = filteredRoomList.get(index.getZeroBased());
-        Room editedRoom = new Room(roomToEdit.getDoctor(), Optional.of(patientBeingServed), false);
-        return new ReversibleActionPairCommand(new NextCommand(roomToEdit, editedRoom, index, patientBeingServed),
-                new UndoNextCommand(editedRoom, roomToEdit, index, patientBeingServed));
+        Room editedRoom = new Room(roomToEdit.getDoctor(), roomToEdit.getCurrentPatient(), true);
+        return new ReversibleActionPairCommand(new BreakCommand(roomToEdit, editedRoom),
+                new ResumeCommand(editedRoom, roomToEdit));
     }
 }
