+++ src/main/java/seedu/address/logic/parser/queue/RemoveRoomCommandParser.java
@@ -1,13 +1,15 @@
-package seedu.address.logic.parser;
+package seedu.address.logic.parser.queue;
 
 import static seedu.address.commons.core.Messages.MESSAGE_INVALID_COMMAND_FORMAT;
 
 import java.util.List;
 
 import seedu.address.commons.core.index.Index;
+import seedu.address.logic.commands.common.ReversibleActionPairCommand;
 import seedu.address.logic.commands.queue.AddConsultationRoomCommand;
 import seedu.address.logic.commands.queue.RemoveRoomCommand;
-import seedu.address.logic.commands.common.ReversibleActionPairCommand;
+import seedu.address.logic.parser.Parser;
+import seedu.address.logic.parser.ParserUtil;
 import seedu.address.logic.parser.exceptions.ParseException;
 import seedu.address.model.Model;
 import seedu.address.model.queue.Room;
