+++ src/main/java/seedu/address/ui/autocomplete/AutoCompleter.java
@@ -7,27 +7,27 @@ import java.util.List;
 import java.util.Map;
 import java.util.Set;
 
-import seedu.address.logic.commands.appointments.AckAppCommand;
-import seedu.address.logic.commands.appointments.AddAppCommand;
-import seedu.address.logic.commands.queue.AddConsultationRoomCommand;
-import seedu.address.logic.commands.appointments.AppointmentsCommand;
-import seedu.address.logic.commands.queue.BreakCommand;
-import seedu.address.logic.commands.appointments.CancelAppCommand;
-import seedu.address.logic.commands.appointments.ChangeAppCommand;
-import seedu.address.logic.commands.queue.DequeueCommand;
-import seedu.address.logic.commands.queue.EnqueueCommand;
 import seedu.address.logic.commands.ExitCommand;
 import seedu.address.logic.commands.HelpCommand;
-import seedu.address.logic.commands.appointments.MissAppCommand;
 import seedu.address.logic.commands.NextCommand;
 import seedu.address.logic.commands.RedoCommand;
-import seedu.address.logic.commands.queue.RemoveRoomCommand;
-import seedu.address.logic.commands.queue.ResumeCommand;
-import seedu.address.logic.commands.appointments.SettleAppCommand;
 import seedu.address.logic.commands.UndoCommand;
+import seedu.address.logic.commands.appointments.AckAppCommand;
+import seedu.address.logic.commands.appointments.AddAppCommand;
+import seedu.address.logic.commands.appointments.AppointmentsCommand;
+import seedu.address.logic.commands.appointments.CancelAppCommand;
+import seedu.address.logic.commands.appointments.ChangeAppCommand;
+import seedu.address.logic.commands.appointments.MissAppCommand;
+import seedu.address.logic.commands.appointments.SettleAppCommand;
 import seedu.address.logic.commands.patients.EditPatientDetailsCommand;
 import seedu.address.logic.commands.patients.ListPatientCommand;
 import seedu.address.logic.commands.patients.RegisterPatientCommand;
+import seedu.address.logic.commands.queue.AddConsultationRoomCommand;
+import seedu.address.logic.commands.queue.BreakCommand;
+import seedu.address.logic.commands.queue.DequeueCommand;
+import seedu.address.logic.commands.queue.EnqueueCommand;
+import seedu.address.logic.commands.queue.RemoveRoomCommand;
+import seedu.address.logic.commands.queue.ResumeCommand;
 import seedu.address.logic.commands.staff.EditStaffDetailsCommand;
 import seedu.address.logic.commands.staff.ListStaffCommand;
 import seedu.address.logic.commands.staff.RegisterStaffCommand;
