+++ src/test/java/seedu/address/logic/parser/AddAppCommandParserTest.java
@@ -6,6 +6,7 @@ import static seedu.address.logic.parser.CommandParserTestUtil.assertParseFailur
 import org.junit.jupiter.api.Test;
 
 import seedu.address.logic.commands.appointments.AddAppCommand;
+import seedu.address.logic.parser.appointments.AddAppCommandParser;
 import seedu.address.model.Model;
 import seedu.address.testutil.TestUtil;
